+++ tests/test_headers.py
@@ -27,10 +27,10 @@ from jsonrpclib.SimpleJSONRPCServer import (
     SimpleJSONRPCRequestHandler,
     SimpleJSONRPCServer,
 )
-from jsonrpclib.utils import from_bytes, to_bytes
+from jsonrpclib.utils import from_bytes
 
 # Tests utilities
-from tests.utilities import UtilityServer
+from tests.utilities import UtilityServer, raw_post
 
 # ------------------------------------------------------------------------------
 
@@ -332,44 +332,14 @@ class HeadersTests(unittest.TestCase):
         self.assertTrue("x-level-2" in headers2)
         self.assertEqual(headers2["x-level-2"], "2")
 
-    def raw_request(self, headers, body=b""):
-        """
-        Sends a raw HTTP request to the test server
-
-        :param headers: The headers of the request, as a string
-        :param body: The raw body of the request
-        :return: The raw answer of the server
-        """
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        sock.settimeout(5)
-        try:
-            sock.connect((HOST, self.port))
-            sock.sendall(
-                to_bytes(
-                    "POST / HTTP/1.0\r\nHost: {0}\r\n{1}\r\n".format(
-                        HOST, headers
-                    )
-                )
-                + body
-            )
-
-            response = b""
-            while True:
-                chunk = sock.recv(4096)
-                if not chunk:
-                    break
-                response += chunk
-        finally:
-            sock.close()
-
-        return response
-
     def test_missing_content_length(self):
         """
         Tests that a request without a Content-Length is rejected with an
         HTTP 411, as its body can't be read
         """
-        response = self.raw_request("Content-Type: application/json-rpc\r\n")
+        response = raw_post(
+            HOST, self.port, "Content-Type: application/json-rpc\r\n"
+        )
         self.assertIn(b"411", response.split(b"\r\n")[0])
 
     def test_invalid_content_length(self):
@@ -378,8 +348,8 @@ class HeadersTests(unittest.TestCase):
         an HTTP 400
         """
         for length in ("abc", "-5", ""):
-            response = self.raw_request(
-                "Content-Length: {0}\r\n".format(length)
+            response = raw_post(
+                HOST, self.port, "Content-Length: {0}\r\n".format(length)
             )
             self.assertIn(
                 b"400",